+++ src/main/java/com/saveapis/inventoryplugin/InventoryPlugin.java
@@ -1,5 +1,7 @@
 package com.saveapis.inventoryplugin;
 
+import com.saveapis.inventoryplugin.commands.InventoryCommand;
+import com.saveapis.inventoryplugin.manager.InventoryManager;
 import com.saveapis.inventoryplugin.manager.UpdateManager;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TextComponent;
@@ -8,22 +10,28 @@ import org.bukkit.Bukkit;
 import org.bukkit.command.ConsoleCommandSender;
 import org.bukkit.plugin.java.JavaPlugin;
 
+import java.util.Objects;
+
 public final class InventoryPlugin extends JavaPlugin {
 
     public static final TextComponent PREFIX = Component.text("[Inventory] ").color(NamedTextColor.BLUE);
-
     public static final ConsoleCommandSender CONSOLE = Bukkit.getConsoleSender();
+    public static InventoryPlugin INSTANCE;
 
     @Override
     public void onLoad() {
+        INSTANCE = this;
         this.saveDefaultConfig();
         UpdateManager.checkForUpdate(this);
+        InventoryManager.initInventory();
         TextComponent message = PREFIX.append(Component.text("Plugin loaded!").color(NamedTextColor.GOLD));
         CONSOLE.sendMessage(message);
     }
 
     @Override
     public void onEnable() {
+        Objects.requireNonNull(this.getCommand("inventory")).setExecutor(new InventoryCommand());
+
         TextComponent message = PREFIX.append(Component.text("Plugin enabled!").color(NamedTextColor.GREEN));
         CONSOLE.sendMessage(message);
     }