+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# InventoryPlugin (v1.3.0)
+# InventoryPlugin (v1.4.0)
 
 ## Fixes
 
@@ -6,11 +6,14 @@ No fixes applied
 
 ## Changes
 
-No changes applied
+ - Register `InventoryCommand` in `plugin.yml`
+ - Register `InventoryCommand` in `InventoryPlugin` Class
+ - Register `InventoryManager.initInventory` in `InventoryPlugin` Class
 
 ## Additions
 
-- Add info messages to `UpdateManager.checkForUpdate`
+- Create `InventoryCommand` Class
+- Create `InventoryManager` Class
 
 ## Deletions